Added likes to posts and comments.
Added like button to posts and comments. Created new field in both post and comment classes which store the number of likes. Set up update method so that whenever a post or message is liked, an update request is sent to the database with the new number of likes. Added sorting of posts and comments to show the one with most likes on top.