Sprint 2 work on posts page and navbar.
Fixed issue where username in Navbar wouldn't upload correctly after registering without refreshing the page. Now it should work fine. Posts and comments now show the name of the user who wrote it, and not the logged in user. Posts and comments now show date and time in which they were made. Only the user who wrote the post or comment can delete it. Fixed issue where a post with comments couldn't be deleted without deleting the comments first, now deleting the posts deletes both the post and its comments.