Not allow same user to like the same post or comment more than once.
Added many to many relationship between post and user and between comment and user to store the users who have liked the post or comment. Now when a user likes a post or comment, they cannot like it again, they can only unlike it, in which case the number of likes goes down.